Converted JARs to OSGi bundles and refactoring
In order to future-proof the reference implementation for the architecture
it is determined that OSGi support should be added to existing JAR
artifacts.  However, OSGi is still not to be used as a technology platform
at the moment.

It is also useful to make sure we depend on the "javax/javaee-api" rather
than the JBoss version when compiling.  This ensure that the JBoss version
of the API does not leak any dependencies into the project.

The unit tests have removed the use of "EasyMock." and "Assert." prefix
in favor of "import static" for the methods to improve readability of the
unit tests.  A Checkstyle rule was also added to detect this.

A sample EasyMock test is provided as well as moving the UtilityClassUtil
test classes to its own package.