Updated dependency versions
Added the use of the versions-maven-plugin to determine which dependency
versions have newer releases.

There are a few that are held back due to M2E support (JAR plugin), JBoss
JavaEE6 due to requiring a dependency not in Central, and Hibernate as
they are Alpha versions.

Report plugins are also sorted.

Findbug Issues are also resolved.