Merged OBR and osgi.tests projects to Assembly
This merges the OBR and osgi.test projects into a new project called
Assembly.  This assembly combines all the projects and dependencies into
one project that is tested using Pax Exam.

It also provides an artifact containing a zipped version of the OBR
containing only the files required for an OBR (i.e. no POMs)